fifth commit

+++ my-app-backend/demo/src/main/java/com/example/demo/Repository/HallsRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.model.HallModel;
 
-public interface HallsRepository extends JpaRepository<HallModel,Integer> {
+public interface HallsRepository extends JpaRepository<HallModel, Long> {
 
     List<HallModel> findByCity(String city);
 }
